+++ src/continuation/aggregate_attractor_fractions.jl
@@ -8,7 +8,7 @@ export aggregate_attractor_fractions
 Aggregate the already-estimated curves of fractions of basins of attraction of similar
 attractors using the same pipeline used by [`GroupingConfig`](@ref).
 The most typical application of this function is to transform the output of
-[`RecurrencesSeededContinuation`](@ref) so that similar attractors, even across parameter
+[`RecurrencesFindAndMatch`](@ref) so that similar attractors, even across parameter
 space, are grouped into one "attractor". Thus, the fractions of their basins are aggregated.
 
 You could also use this function to aggregate attractors and their fractions even in
@@ -28,7 +28,7 @@ in [Extinction of a species in a multistable competition model](@ref).
 1. `fractions_curves`: a vector of dictionaries mapping labels to basin fractions.
 2. `attractors_info`: a vector of dictionaries mapping labels to attractors.
    1st and 2nd argument are exactly like the return values of
-   [`continuation`](@ref) with [`RecurrencesSeededContinuation`](@ref)
+   [`continuation`](@ref) with [`RecurrencesFindAndMatch`](@ref)
    (or, they can be the return of [`basins_fractions`](@ref)).
 3. `featurizer`: a 1-argument function to map an attractor into an appropriate feature
    to be grouped later. Features expected by [`GroupingConfig`](@ref) are `SVector`.